% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{calc_ts_np}
\alias{calc_ts_np}
\title{Calculate time series frequency and proportion}
\usage{
calc_ts_np(data, varname, time)
}
\arguments{
\item{data}{A data frame}

\item{varname}{Variable to compute}

\item{time}{Time variable}
}
\value{
A tibble with 4 columns

* `time`: the time variable

* `varname`: the variable to compute

* `n`: frequency of each category of `varname` per time step

* `prop`: proportion of each category of `varname` per time step
}
\description{
Calculate time series frequency and proportion
}
\examples{
df <- patients
calc_ts_np(data = df, varname = age_class, time = onset_month)
}
\author{
Thinh Ong \url{https://github.com/thinhong}
}
