% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rds_w_increment.R
\name{write_rds_w_increment}
\alias{write_rds_w_increment}
\title{Write object to RDS file with incrementing number suffixed to file name}
\usage{
write_rds_w_increment(object, dir, fname)
}
\arguments{
\item{object}{object to be written}

\item{dir}{directory for the file}

\item{fname}{desired file name}
}
\value{
a string confirming where the file has been written to
}
\description{
Write object to RDS file with incrementing number suffixed to file name
}
\examples{
write_rds_w_increment(identity, "~", "temp")
# "New file written at ~/temp.rds"
write_rds_w_increment(identity, "~", "temp")
# "New file written at ~/temp_01.rds"
write_rds_w_increment(identity, "~", "temp")
# "New file written at ~/temp_02.rds"
}
\author{
Tuyen Huynh \url{https://github.com/tuyenhn}
}
