% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_ts_stacked_bar}
\alias{plot_ts_stacked_bar}
\title{Plot time series stacked bar chart}
\usage{
plot_ts_stacked_bar(data, varname, time, style = F)
}
\arguments{
\item{data}{A data frame}

\item{varname}{Variable to plot}

\item{time}{Time variable}

\item{style}{Add minimal theme and ggsci palette}
}
\value{
A list contains:

* `data`: A data frame created by `calc_ts_np()` using to plot

* `plot`: A ggplot object for the time series stacked bar chart
}
\description{
Plot time series stacked bar chart
}
\examples{
df <- patients
outp <- plot_ts_stacked_bar(data = df, varname = age_class, time = onset_month)
outp$plot
# Add other ggplot elements
outp$plot + ggplot2::theme_minimal()
}
\author{
Thinh Ong \url{https://github.com/thinhong}
}
